+++ packages/expo-modules-core/android/src/main/java/expo/modules/kotlin/ModuleHolder.kt
@@ -1,23 +1,24 @@
 package expo.modules.kotlin
 
 import com.facebook.react.bridge.ReadableArray
-import expo.modules.core.utilities.ifNull
 import expo.modules.kotlin.events.BasicEventListener
 import expo.modules.kotlin.events.EventListenerWithPayload
 import expo.modules.kotlin.events.EventListenerWithSenderAndPayload
 import expo.modules.kotlin.events.EventName
+import expo.modules.kotlin.exception.FunctionCallException
 import expo.modules.kotlin.exception.MethodNotFoundException
+import expo.modules.kotlin.exception.exceptionDecorator
 import expo.modules.kotlin.modules.Module
 
 class ModuleHolder(val module: Module) {
   val definition = module.definition()
   val name get() = definition.name
 
-  fun call(methodName: String, args: ReadableArray, promise: Promise) {
-    val method = definition.methods[methodName].ifNull {
-      promise.reject(MethodNotFoundException(methodName, definition.name))
-      return
-    }
+  fun call(methodName: String, args: ReadableArray, promise: Promise) = exceptionDecorator({
+    FunctionCallException(methodName, definition.name, it)
+  }) {
+    val method = definition.methods[methodName]
+      ?: throw MethodNotFoundException()
 
     method.call(args, promise)
   }