+++ packages/expo-modules-core/android/src/main/java/expo/modules/kotlin/methods/AnyMethod.kt
@@ -2,31 +2,26 @@ package expo.modules.kotlin.methods
 
 import com.facebook.react.bridge.ReadableArray
 import expo.modules.kotlin.Promise
+import expo.modules.kotlin.exception.ArgumentCastException
 import expo.modules.kotlin.exception.CodedException
 import expo.modules.kotlin.exception.InvalidArgsNumberException
-import expo.modules.kotlin.exception.UnexpectedException
+import expo.modules.kotlin.exception.exceptionDecorator
 import expo.modules.kotlin.iterator
 import expo.modules.kotlin.recycle
 import expo.modules.kotlin.types.AnyType
-import kotlin.jvm.Throws
 
 abstract class AnyMethod(
   protected val name: String,
   private val desiredArgsTypes: Array<AnyType>
 ) {
+  @Throws(CodedException::class)
   fun call(args: ReadableArray, promise: Promise) {
     if (desiredArgsTypes.size != args.size()) {
-      promise.reject(InvalidArgsNumberException(args.size(), desiredArgsTypes.size))
-      return
+      throw InvalidArgsNumberException(args.size(), desiredArgsTypes.size)
     }
-    try {
-      val convertedArgs = convertArgs(args)
-      callImplementation(convertedArgs, promise)
-    } catch (codedError: CodedException) {
-      promise.reject(codedError)
-    } catch (e: Throwable) {
-      promise.reject(UnexpectedException(e))
-    }
+
+    val convertedArgs = convertArgs(args)
+    callImplementation(convertedArgs, promise)
   }
 
   @Throws(CodedException::class)
@@ -40,9 +35,13 @@ abstract class AnyMethod(
     val argIterator = args.iterator()
     desiredArgsTypes
       .withIndex()
-      .forEach { (index, type) ->
+      .forEach { (index, desiredType) ->
         argIterator.next().recycle {
-          finalArgs[index] = type.convert(this)
+          exceptionDecorator({ cause ->
+            ArgumentCastException(desiredType.kType, index, type, cause)
+          }) {
+            finalArgs[index] = desiredType.convert(this)
+          }
         }
       }
     return finalArgs