+++ packages/expo-modules-core/ios/Swift/Arguments/Convertibles.swift
@@ -19,7 +19,7 @@ extension URL: ConvertibleArgument {
       // If it has no scheme, we assume it was the file path.
       return url.scheme != nil ? url : URL(fileURLWithPath: uri)
     }
-    throw Conversions.ConvertingError<URL>(value: value)
+    throw Conversions.ConvertingException<URL>(value)
   }
 }
 
@@ -37,7 +37,7 @@ extension UIColor: ConvertibleArgument {
     if let value = value as? Int {
       return try Conversions.toColor(argb: UInt64(value)) as! Self
     }
-    throw Conversions.ConvertingError<UIColor>(value: value)
+    throw Conversions.ConvertingException<UIColor>(value)
   }
 }
 
@@ -52,7 +52,7 @@ extension CGPoint: ConvertibleArgument {
       let args = try Conversions.pickValues(from: value, byKeys: ["x", "y"], as: Double.self)
       return CGPoint(x: args[0], y: args[1])
     }
-    throw Conversions.ConvertingError<CGPoint>(value: value)
+    throw Conversions.ConvertingException<CGPoint>(value)
   }
 }
 
@@ -65,7 +65,7 @@ extension CGSize: ConvertibleArgument {
       let args = try Conversions.pickValues(from: value, byKeys: ["width", "height"], as: Double.self)
       return CGSize(width: args[0], height: args[1])
     }
-    throw Conversions.ConvertingError<CGSize>(value: value)
+    throw Conversions.ConvertingException<CGSize>(value)
   }
 }
 
@@ -78,7 +78,7 @@ extension CGVector: ConvertibleArgument {
       let args = try Conversions.pickValues(from: value, byKeys: ["dx", "dy"], as: Double.self)
       return CGVector(dx: args[0], dy: args[1])
     }
-    throw Conversions.ConvertingError<CGVector>(value: value)
+    throw Conversions.ConvertingException<CGVector>(value)
   }
 }
 
@@ -91,7 +91,7 @@ extension CGRect: ConvertibleArgument {
       let args = try Conversions.pickValues(from: value, byKeys: ["x", "y", "width", "height"], as: Double.self)
       return CGRect(x: args[0], y: args[1], width: args[2], height: args[3])
     }
-    throw Conversions.ConvertingError<CGRect>(value: value)
+    throw Conversions.ConvertingException<CGRect>(value)
   }
 }
 
@@ -99,9 +99,9 @@ extension CGColor: ConvertibleArgument {
   public static func convert(from value: Any?) throws -> Self {
     do {
       return try UIColor.convert(from: value).cgColor as! Self
-    } catch _ as Conversions.ConvertingError<UIColor> {
+    } catch _ as Conversions.ConvertingException<UIColor> {
       // Rethrow `ConvertingError` with proper type
-      throw Conversions.ConvertingError<CGColor>(value: value)
+      throw Conversions.ConvertingException<CGColor>(value)
     }
   }
 }