+++ packages/expo-modules-core/ios/Swift/Functions/ConcreteFunction.swift
@@ -15,6 +15,8 @@ public final class ConcreteFunction<Args, ReturnType>: AnyFunction {
 
   public var queue: DispatchQueue?
 
+  public var isAsync: Bool = true
+
   let closure: ClosureType
 
   let argTypes: [AnyArgumentType]
@@ -45,8 +47,8 @@ public final class ConcreteFunction<Args, ReturnType>: AnyFunction {
     } catch let error as CodedError {
       promise.reject(FunctionCallException(name).causedBy(error))
       return
-    } catch let error {
-      promise.reject(UnexpectedError(error))
+    } catch {
+      promise.reject(UnexpectedException(error))
       return
     }
     if !takesPromise {
@@ -84,6 +86,11 @@ public final class ConcreteFunction<Args, ReturnType>: AnyFunction {
     return self
   }
 
+  public func runSynchronously() -> Self {
+    self.isAsync = false
+    return self
+  }
+
   private func argumentType(atIndex index: Int) -> AnyArgumentType? {
     return (0..<argTypes.count).contains(index) ? argTypes[index] : nil
   }
@@ -107,18 +114,18 @@ public final class ConcreteFunction<Args, ReturnType>: AnyFunction {
 
 internal class InvalidArgsNumberException: GenericException<(received: Int, expected: Int)> {
   override var reason: String {
-    "Received \(params.received) arguments, but \(params.expected) was expected."
+    "Received \(param.received) arguments, but \(param.expected) was expected"
   }
 }
 
 internal class ArgumentCastException: GenericException<(index: Int, type: AnyArgumentType)> {
   override var reason: String {
-    "Argument at index '\(params.index)' couldn't be casted to type '\(params.type.description)'."
+    "Argument at index '\(param.index)' couldn't be cast to type \(param.type.description)"
   }
 }
 
 internal class FunctionCallException: GenericException<String> {
   override var reason: String {
-    "Call to function '\(params)' has been rejected."
+    "Call to function '\(param)' has been rejected"
   }
 }