+++ packages/expo/cli/utils/prompts.ts
@@ -1,4 +1,4 @@
-import prompts, { Options, PromptObject, PromptType } from 'prompts';
+import prompts, { Options, Choice, PromptObject, PromptType } from 'prompts';
 
 import { CI } from './env';
 import { AbortCommandError, CommandError } from './errors';
@@ -6,6 +6,10 @@ import { AbortCommandError, CommandError } from './errors';
 export type Question<V extends string = string> = PromptObject<V> & {
   optionsPerPage?: number;
 };
+
+export interface ExpoChoice<T> extends Choice {
+  value: T;
+}
 
 export { PromptType };
 
@@ -17,7 +21,7 @@ export default async function prompt(
 ) {
   questions = Array.isArray(questions) ? questions : [questions];
   if (CI && questions.length !== 0) {
-    let message = `Input is required, but Expo CLI is in non-interactive mode.\n`;
+    let message = `Input is required, but 'npx expo' is in non-interactive mode.\n`;
     if (nonInteractiveHelp) {
       message += nonInteractiveHelp;
     } else {
@@ -105,3 +109,23 @@ export function resumeInteractions(options: Omit<InteractionOptions, 'pause'> = 
     listener({ pause: false, ...options });
   }
 }
+
+/** Select an option from a list of options. */
+export async function selectAsync<T>(
+  message: string,
+  choices: ExpoChoice<T>[],
+  options?: PromptOptions
+): Promise<T> {
+  const { value } = await prompt(
+    {
+      message,
+      choices,
+      name: 'value',
+      type: 'select',
+    },
+    options
+  );
+  return value ?? null;
+}
+
+export const promptAsync = prompt;