+++ packages/expo-modules-core/CHANGELOG.md
@@ -20,6 +20,7 @@
 ### 🐛 Bug fixes
 
 - It's no longer possible to directly call methods from the `ModuleDefinition` in the `ViewManagers` on Android. ([#15741](https://github.com/expo/expo/pull/15741) by [@lukmccall](https://github.com/lukmccall))
+- Fix compatibility with react-native 0.66. ([#15914](https://github.com/expo/expo/pull/15914) by [@kudo](https://github.com/kudo))
 
 ### 💡 Others